requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
--index-url https://mirrors.aliyun.com/pypi/simple/
--trusted-host mirrors.aliyun.com

aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via
    #   dashscope
    #   langchain-community
    #   qianfan
aiolimiter==1.2.1
    # via qianfan
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via typer
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   openai
async-timeout==4.0.3
    # via
    #   aiohttp
    #   langchain-classic
attrs==25.4.0
    # via aiohttp
authlib==1.6.9
    # via weaviate-client
bce-python-sdk==0.9.63
    # via qianfan
beautifulsoup4==4.14.3
    # via wikipedia
cachetools==7.0.5
    # via qianfan
certifi==2026.2.25
    # via
    #   dashscope
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.5
    # via requests
click==8.3.1
    # via
    #   duckduckgo-search
    #   typer
colorama==0.4.6
    # via
    #   click
    #   tqdm
cryptography==46.0.5
    # via
    #   authlib
    #   dashscope
dashscope==1.25.13
    # via -r requirements.in
dataclasses-json==0.6.7
    # via langchain-community
deprecation==2.1.0
    # via weaviate-client
dill==0.4.1
    # via multiprocess
diskcache==5.6.3
    # via qianfan
distro==1.9.0
    # via openai
duckduckgo-search==8.1.1
    # via -r requirements.in
exceptiongroup==1.3.1
    # via anyio
filelock==3.25.1
    # via huggingface-hub
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2026.2.0
    # via huggingface-hub
future==1.0.0
    # via bce-python-sdk
greenlet==3.3.2
    # via sqlalchemy
grpcio==1.78.0
    # via weaviate-client
h11==0.16.0
    # via httpcore
hf-xet==1.3.2
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   huggingface-hub
    #   langgraph-sdk
    #   langsmith
    #   ollama
    #   openai
    #   weaviate-client
httpx-sse==0.4.3
    # via langchain-community
huggingface-hub==1.6.0
    # via
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jieba==0.42.1
    # via -r requirements.in
jiter==0.13.0
    # via openai
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==1.2.11
    # via -r requirements.in
langchain-classic==1.0.2
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.4.1
    # via -r requirements.in
langchain-core==1.2.18
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-classic
    #   langchain-community
    #   langchain-ollama
    #   langchain-openai
    #   langchain-text-splitters
    #   langchain-weaviate
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-ollama==1.0.1
    # via -r requirements.in
langchain-openai==1.1.11
    # via -r requirements.in
langchain-text-splitters==1.1.1
    # via
    #   -r requirements.in
    #   langchain-classic
langchain-weaviate==0.0.6
    # via -r requirements.in
langgraph==1.1.0
    # via
    #   -r requirements.in
    #   langchain
langgraph-checkpoint==4.0.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==1.0.8
    # via langgraph
langgraph-sdk==0.3.10
    # via langgraph
langsmith==0.7.16
    # via
    #   langchain-classic
    #   langchain-community
    #   langchain-core
lxml==6.0.2
    # via duckduckgo-search
markdown-it-py==4.0.0
    # via rich
marshmallow==3.26.2
    # via dataclasses-json
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.19
    # via qianfan
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.2.6
    # via
    #   langchain-community
    #   langchain-weaviate
    #   pandas
    #   transformers
ollama==0.6.1
    # via langchain-ollama
openai==2.26.0
    # via langchain-openai
orjson==3.11.7
    # via
    #   langgraph-sdk
    #   langsmith
ormsgpack==1.12.2
    # via langgraph-checkpoint
packaging==26.0
    # via
    #   deprecation
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   marshmallow
    #   transformers
pandas==2.3.3
    # via -r requirements.in
primp==1.1.2
    # via duckduckgo-search
prompt-toolkit==3.0.52
    # via qianfan
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
protobuf==6.33.5
    # via weaviate-client
pycparser==3.0
    # via cffi
pycryptodome==3.23.0
    # via bce-python-sdk
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-classic
    #   langchain-core
    #   langgraph
    #   langsmith
    #   ollama
    #   openai
    #   pydantic-settings
    #   qianfan
    #   weaviate-client
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via langchain-community
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   qianfan
pytz==2026.1.post1
    # via pandas
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   langchain-classic
    #   langchain-community
    #   langchain-core
    #   qianfan
    #   transformers
qianfan==0.4.12.3
    # via -r requirements.in
regex==2026.2.28
    # via
    #   tiktoken
    #   transformers
requests==2.32.5
    # via
    #   -r requirements.in
    #   dashscope
    #   langchain-classic
    #   langchain-community
    #   langsmith
    #   qianfan
    #   requests-toolbelt
    #   tiktoken
    #   wikipedia
requests-toolbelt==1.0.0
    # via langsmith
rich==14.3.3
    # via
    #   qianfan
    #   typer
safetensors==0.7.0
    # via transformers
shellingham==1.5.4
    # via typer
simsimd==6.5.16
    # via langchain-weaviate
six==1.17.0
    # via
    #   bce-python-sdk
    #   python-dateutil
sniffio==1.3.1
    # via openai
soupsieve==2.8.3
    # via beautifulsoup4
sqlalchemy==2.0.48
    # via
    #   langchain-classic
    #   langchain-community
tenacity==8.5.0
    # via
    #   langchain-community
    #   langchain-core
    #   qianfan
tiktoken==0.12.0
    # via langchain-openai
tokenizers==0.22.2
    # via transformers
tqdm==4.67.3
    # via
    #   huggingface-hub
    #   openai
    #   transformers
transformers==5.3.0
    # via -r requirements.in
typer==0.24.1
    # via
    #   huggingface-hub
    #   qianfan
    #   transformers
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   beautifulsoup4
    #   cryptography
    #   exceptiongroup
    #   grpcio
    #   huggingface-hub
    #   langchain-core
    #   multidict
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
uuid-utils==0.14.1
    # via
    #   langchain-core
    #   langsmith
validators==0.35.0
    # via weaviate-client
wcwidth==0.6.0
    # via prompt-toolkit
weaviate-client==4.20.4
    # via langchain-weaviate
websocket-client==1.9.0
    # via dashscope
wikipedia==1.4.0
    # via -r requirements.in
xxhash==3.6.0
    # via
    #   langgraph
    #   langsmith
yarl==1.23.0
    # via aiohttp
zstandard==0.25.0
    # via langsmith
